requirements.txt
requests==2.31.0
python-dotenv==1.0.1

#pip install -r requirements.txt    to install all dependencies